requirements.txt
appdirs==1.4.4
asgiref==3.3.4
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.7
click==8.0.1
colorama==0.4.4
coverage==5.5
Django==3.2.4
django-cors-headers==3.7.0
django-environ==0.4.5
django-extensions==3.1.3
django-filter==2.4.0
djangorestframework==3.12.4
flake8==3.9.2
iniconfig==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.21.0
packaging==21.0
pathspec==0.8.1
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pydot==1.4.2
pyflakes==2.3.1
pyparsing==2.4.7
pytest==6.2.4
pytz==2021.1
PyYAML==5.4.1
regex==2021.7.6
sqlparse==0.4.1
toml==0.10.2
tomli==1.2.0
uritemplate==3.0.1
